% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(key1, key2)
}
\arguments{
\item{key1}{Chave real de encriptação do texto.}

\item{key2}{Chave a ser comparada.}
}
\value{
A acurácia entre as duas chaves.
}
\description{
Função que calcula a acurácia entre duas chaves.
}
\details{
A acurácia entre duas chaves é definida como a proporção de vezes que os carracteres
na mesma posição do texto são iguais nas duas chaves.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
