% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text}
\usage{
clean_text(text, accent = FALSE)
}
\arguments{
\item{text}{Texto a ser limpado}

\item{accent}{logical. Se FALSE a função tira os acentos do texto. Se TRUE, ela deixa. Default é
tirar os acentos.}
}
\value{
O texto limpo.
}
\description{
Limpa um texto.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
