% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_subs.R
\name{decode_subs}
\alias{decode_subs}
\title{Decode Text}
\usage{
decode_subs(key, text_str, alpha = c(LETTERS, " "))
}
\arguments{
\item{key}{vetor com a chave.}

\item{text_str}{string de texto.}

\item{alpha}{vetor com o alfabeto. Default é as 26 letras do alfabeto e espaço.}
}
\value{
Retorna vetor com a string decriptada.
}
\description{
Uma função que decripta um texto.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
