% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_mcmc.R
\name{decrypt_mcmc}
\alias{decrypt_mcmc}
\title{Decrypt Substitution Cipher With MCMC}
\usage{
decrypt_mcmc(
  encoded,
  key0,
  reference_bigrams,
  reference_freq,
  pr.method = 1,
  nr = 10000,
  p = 1,
  alpha = c(LETTERS, " ")
)
}
\arguments{
\item{encoded}{texto encriptado.}

\item{key0}{chave inicial.}

\item{reference_bigrams}{vetor de bigramas do texto de referência.}

\item{reference_freq}{vetor de frequência dos bigramas do texto de referência.}

\item{pr.method}{= 1, nr = 10000, p = 1, alpha = c(LETTERS, ' ')}

\item{nr}{número inteiro que indica o número de iterações do algoritmo. Default é 10.000.}

\item{p}{número inteiro que indica o valor do parâmetro de escala. Default é 1.}

\item{alpha}{vetor com o alfabeto. Default é as 26 letras do alfabeto e espaço.}
}
\value{
Retorna uma lista com o argmax da função escore, a respectiva chave e a
quantidade de vezes que a proposta foi aceita.
}
\description{
Realizada o Markov Chain Monte Carlo (MCMC) para cifras de substituição.
}
\references{
CHEN,J.J. & ROSENTHAL,J.(2012). Decrypting classical cipher text using markov chain monte carlo.Statisticsand Computing22, 397–413.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
