% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_p.R
\name{funcao_p}
\alias{funcao_p}
\title{Score Function}
\usage{
funcao_p(encoded_text, reference_bigrams, reference_freq, pipower = 0.25)
}
\arguments{
\item{encoded_text}{texto encriptado.}

\item{reference_bigrams}{Vetor de bigramas do texto de referência.}

\item{reference_freq}{vetor de frequência dos bigramas do texto de referência.}

\item{pipower}{multiplicador para suavizar a função escore e evitar erros numéricos.
Default é 0.25, o mesmo de Chen & Rosenthal (2012).}
}
\value{
Retorna o valor da função escore dado os argumentos fornecidos.
}
\description{
Função que calcula o valor da função escore baseado no texto encriptado.
}
\references{
CHEN,J.J. & ROSENTHAL,J.(2012). Decrypting classical cipher text using markov chain monte carlo.Statisticsand Computing22, 397–413.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
