% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bigrams_freq.R
\name{get_bigrams_freq}
\alias{get_bigrams_freq}
\title{Get Bigrams Frequency}
\usage{
get_bigrams_freq(bigrams)
}
\arguments{
\item{bigrams}{vetor de bigramas.}
}
\value{
data frame de frequências.
}
\description{
Funcao para gerar a tabela de frequência dos bigramas.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
