% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bigrams_from_text.R
\name{get_bigrams_from_text}
\alias{get_bigrams_from_text}
\title{Get Bigrams From Text}
\usage{
get_bigrams_from_text(text_vector)
}
\arguments{
\item{text_vector}{Vetor de strings ou string.}
}
\value{
Vetor de bigramas.
}
\description{
Cria bigramas de caracteres a partir do texto.
}
\author{
Taís Bellini e Juliana Sena de Souza
}
