% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{Get Text}
\usage{
get_text(path)
}
\arguments{
\item{path}{o nome do local e arquivo do qual os dados devem ser lidos.}
}
\value{
O arquivo de texto.
}
\description{
Uma função que lê o arquivo de texto em UTF-8.
}
\seealso{
\code{\link[base]{readLines}}
}
\author{
Taís Bellini e Juliana Sena de Souza
}
